Restructored application file structure and refactored widgets to better align with readings from flutterarbiter.com
