+++ lib/features/install_ffmpeg/verify_install.dart
@@ -3,9 +3,9 @@ import 'dart:io';
 import 'dart:isolate';
 
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:verticrowlabs_media_converter/ffmpeg_install_helper/ffmpeg_install_helper.dart';
-import 'package:verticrowlabs_media_converter/utils/common_variables.dart';
-import 'package:verticrowlabs_media_converter/utils/pwsh_cmd.dart';
+import 'package:verticrowlabs_media_converter/features/install_ffmpeg/install_ffmpeg.dart';
+import 'package:verticrowlabs_media_converter/infrastructure/common_variables/common_variables.dart';
+import 'package:verticrowlabs_media_converter/infrastructure/common_variables/pwsh_cmd.dart';
 
 //TO-DO: #18 add ffmpeg installation ability for linux and macos. @anadreau
 //TO-DO: #19 implement ffmpeg command depending on OS. @anadreau
