+++ lib/features/thumbnail_generator/thumbnail_widget.dart
@@ -2,7 +2,7 @@ import 'dart:io';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:verticrowlabs_media_converter/thumbnail_generator/thumbnail_barrel.dart';
+import 'package:verticrowlabs_media_converter/features/thumbnail_generator/thumbnail_barrel.dart';
 
 //Thumbnail function goes here.
 ///[ConsumerWidget] to return the thumbnail of the media file that will be
