+++ lib/infrastructure/common_variables/common_variables.dart
