+++ lib/infrastructure/common_variables/ffmpeg_cmd.dart
@@ -5,9 +5,9 @@
 // -c:v libx264 output.mp4
 
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:verticrowlabs_media_converter/file_parsing/file_parsing_barrel.dart';
-import 'package:verticrowlabs_media_converter/media_conversion/media_conversion_barrel.dart';
-import 'package:verticrowlabs_media_converter/utils/time_selector.dart';
+import 'package:verticrowlabs_media_converter/features/file_parsing/file_parsing_barrel.dart';
+import 'package:verticrowlabs_media_converter/features/media_conversion/media_conversion_barrel.dart';
+import 'package:verticrowlabs_media_converter/infrastructure/common_widgets/time_selector.dart';
 
 ///[Provider] that handles logic for which ffmpeg cmd is selected when
 ///convert_button is pressed.
