+++ lib/infrastructure/common_variables/pwsh_cmd.dart
