+++ lib/infrastructure/common_widgets/file_selector_widget.dart
@@ -1,12 +1,12 @@
 import 'package:file_picker/file_picker.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:verticrowlabs_media_converter/file_parsing/file_parsing_barrel.dart';
-import 'package:verticrowlabs_media_converter/media_conversion/conversion_status.dart';
-import 'package:verticrowlabs_media_converter/media_snipping/media_snipping.dart';
-import 'package:verticrowlabs_media_converter/thumbnail_generator/thumbnail_barrel.dart';
-import 'package:verticrowlabs_media_converter/utils/common_variables.dart';
-import 'package:verticrowlabs_media_converter/utils/time_selector.dart';
+import 'package:verticrowlabs_media_converter/features/file_parsing/file_parsing_barrel.dart';
+import 'package:verticrowlabs_media_converter/features/media_conversion/conversion_status.dart';
+import 'package:verticrowlabs_media_converter/features/media_snipping/media_snipping.dart';
+import 'package:verticrowlabs_media_converter/features/thumbnail_generator/thumbnail_barrel.dart';
+import 'package:verticrowlabs_media_converter/infrastructure/common_variables/common_variables.dart';
+import 'package:verticrowlabs_media_converter/infrastructure/common_widgets/time_selector.dart';
 
 ///[ConsumerWidget] that sets [fileInputStringProvider] when button is pressed.
 class FileSelector extends ConsumerWidget {
