+++ lib/infrastructure/common_widgets/file_type_drop_down.dart
@@ -2,8 +2,8 @@ import 'dart:developer';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:verticrowlabs_media_converter/media_conversion/container_type.dart';
-import 'package:verticrowlabs_media_converter/utils/common_variables.dart';
+import 'package:verticrowlabs_media_converter/features/media_conversion/container_type.dart';
+import 'package:verticrowlabs_media_converter/infrastructure/common_variables/common_variables.dart';
 
 ///List of MediaContainerType as enum
 const List<MediaContainerType> dropDownListEnum = MediaContainerType.values;
