+++ lib/infrastructure/common_widgets/resolution_drop_down.dart
@@ -2,8 +2,8 @@ import 'dart:developer';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:verticrowlabs_media_converter/media_conversion/media_conversion_barrel.dart';
-import 'package:verticrowlabs_media_converter/utils/common_variables.dart';
+import 'package:verticrowlabs_media_converter/features/media_conversion/media_conversion_barrel.dart';
+import 'package:verticrowlabs_media_converter/infrastructure/common_variables/common_variables.dart';
 
 ///[List] of [MediaScale] values for use in [MediaDropDown]
 final List<MediaScale> mediaScaleDropDownList = MediaScale.values.toList();
